Integrated graphql with typeORM/database, can now add user and search for all users.
