+++ src/index.ts
@@ -1,20 +1,6 @@
 import { AppDataSource } from './data-source';
-import { User } from './entity/User';
+import { initApolloServer } from './apolloServer';
 
 AppDataSource.initialize()
-  .then(async () => {
-    console.log('Inserting a new user into the database...');
-    const user = new User();
-    user.firstName = 'Timber';
-    user.lastName = 'Saw';
-    user.age = 25;
-    await AppDataSource.manager.save(user);
-    console.log('Saved a new user with id: ' + user.id);
-
-    console.log('Loading users from the database...');
-    const users = await AppDataSource.manager.find(User);
-    console.log('Loaded users: ', users);
-
-    console.log('Here you can setup and run express / fastify / any other framework.');
-  })
+  .then(initApolloServer)
   .catch((error) => console.log(error));
