+++ src/schema.graphql
@@ -1,0 +1,15 @@
+type Query {
+  users: [User]
+}
+
+type Mutation {
+  createUser(name: String!, email: String!, password: String!, birthdate: String!): User!
+}
+
+type User {
+  id: String!
+  name: String!
+  email: String!
+  password: String!
+  birthdate: String!
+}
